Trying to Implement Random Forest
In randomForest, I implement OneHotEncoding using Pandas. One hot encoding is based on sequence encoding. Using this I was able to start trying to implement the randomForest prediction algorithm. I am currently struggling with the features and the labels as I am having a slightly hard time differentiating them in such a way that it would be able to progress through the randomForest algorithm. The step I'm currently stuck on helps me separate the data with a 80/20 split of training and testing. Besides this, I have just updated and moved around a few of the old files for organizational purposes but that should not be a major problem